requirements.txt
torch==1.1.0
scipy==1.5.4
torchvision==0.3.0
seaborn==0.11.0
numpy==1.19.4
faiss==1.6.4
Pillow==7.1.2
scikit_learn==0.21.2
networkx
